using sdint values instead of raw c

+++ include/kernel/multitasking/task.h
@@ -15,18 +15,18 @@ struct Task
 	struct Task *next_task;
 	int state; /* (running, waiting, ready)*/
 
-	uint32_t eax, ebx, ecx, edx;
-	uint32_t esi, edi, ebp;
-	uint32_t esp;			   /* Stack pointer*/
-	uint32_t eip;			   /* Instruction pointer*/
-	uint32_t eflags;			/* Flags register*/
+	unsigned long eax, ebx, ecx, edx;
+	unsigned long esi, edi, ebp;
+	unsigned long esp;			   /* Stack pointer*/
+	unsigned long eip;			   /* Instruction pointer*/
+	unsigned long eflags;			/* Flags register*/
 
 	/* Optional fields*/
 	int scheduling_policy;  /* Task's scheduling policy*/
 	int scheduling_priority;/* Task's scheduling priority*/
 	int pid;				/* Process ID the task belongs to*/
 	char task_name[64];		/* Task name (for debugging/monitoring purposes)*/
-	uint32_t cpu_time_consumed; /* CPU time consumed by the task so far*/
+	unsigned long cpu_time_consumed; /* CPU time consumed by the task so far*/
 };
 
 int create_task(struct task *task, void (*entry_point)(void));
